requirements.txt
setuptools
cython>=3.0.10
kivy>=2.1.0
plyer>=2.1.0
pyjnius
buildozer==1.5.0
python-for-android==2024.1.21